support float width and hight for the ellipse

+++ src/internal_classes.py
@@ -36,8 +36,8 @@ class Shape:
     type_id: int
     x: int
     y: int
-    w: int
-    h: int
+    w: float
+    h: float
     rot_deg: int
     color: Color
     is_mask: bool
